+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stoker-platform/utils
 
+## 0.5.83
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.68
+
 ## 0.5.82
 
 ### Patch Changes